pytests: Audio loop does not work if device have only speaker
If a device has only a speaker, and does not have any other output
devices, then the audio loop will fail. The reason is, it will try
to disable the Headphone (if we're choosing to output from speaker),
and enable the Speaker. However, since the device does not have
Headphone at all, that call will raise an exception (as expected).

BUG=b:74299033
TEST=Manually run on a device with only speaker, and watch the behavior.

Change-Id: Id87966a538886bd53898aece649f84ce7847a684
Reviewed-on: https://chromium-review.googlesource.com/960184
Commit-Ready: Shen-En Shih <petershih@chromium.org>
Tested-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Yong Hong <yhong@google.com>