station-based test list: clear state on start
For station-based test lists, when we restart Goofy (by rebooting or
factory_restart), we should always start from the beginning of the test
list.

Note that this will break the test lists containing "Reboot" test.
However, this should not happen for most station-based test list.  We
can discuss how to deal with it when we need this.

BUG=None
TEST=make test

Change-Id: I40bb71d9e27bbd132b679df4d1382a063cdea245
Reviewed-on: https://chromium-review.googlesource.com/961995
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Yong Hong <yhong@google.com>