Instalog: Fix race condition and TearDown in buffer plugins
If the buffer is producing the first event and the consumer try to
restore metadata, the consumer will think the buffer is broken.
There's some data in the buffer, but there's no metadata is saved yet.

Use the flag "sync" when stopping buffer plugin.

Modify the Next() method in buffer priority plugin to speed up the
plugin.

BUG=b:67661968
TEST=manually test on local device

Change-Id: I5453032801e46f666cd5e8cda3e2235cf2be6bb7
Reviewed-on: https://chromium-review.googlesource.com/970084
Commit-Ready: Chun-Tsen Kuo <chuntsen@chromium.org>
Tested-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>