probe: Implements the probe function for `cros-camera-tool`.
This CL implements the probe function which parses the output of
`cros-camera-tool` to obtain the identities of MIPI cameras.

This CL also adds the new probe function into the file
`py/hwid/v3/default_probe_statement.json` for a temporary solution
to make sure that MIPI cameras is probed when the HWID string is
generated for all projects.

BUG=b:70357756
TEST=make test; manually test on DUT.

Change-Id: I147918ca0eb3fd31815b895e2496701a5f7d8aea
Reviewed-on: https://chromium-review.googlesource.com/970181
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>