utils: Move FirmwareImage from gooftool.crosfw to utils.fmap.
The crosfw.FirmwareImage is a wrapper for accessing FMAP based firmware
image easily. The only dependency of it is fmap. More tools can use it
to easily access FMAP firmware images (not just Chrome OS firmware
images) if it's in fmap module instead of gooftool (which has lots of
extra dependency).

As a result, we should move FirmareImage to cros.factory.utils.fmap
so more programs can use it directly.

BUG=None
TEST=make test

Change-Id: I10e39c56ddb6605bcb97be82eaaf32ee31c357b3
Reviewed-on: https://chromium-review.googlesource.com/1007462
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>