pygpt: Extend named tuple objects into more useful objects with properties.
The raw implementation of pygpt.GPT is exposing objects in pure named
tuples, which makes it pretty hard to use since every pygpt user (for
example image_tool) has to implement the logic of LBA calculation,
encoding and decoding of UTF-16 strings, GUID conversion, ... etc.

This change tries to enhance the named tuple classes by extending
helpful properties in derived classes, making it more similar to the
Partition object defined in image_tool.

Additionally, as discovered in image_tool.Partition, the GPT objects may
need more information that is not available in GPT, for instance disk
image file name or block size. This is solved by the __dict__ in
GPTObject objects and maintained in Clone and ReadFrom.

With all these changes, we can now largely simplify the manipulation of
GPT objects using the help functions and properties.

BUG=chromium:834237
TEST=make test

Change-Id: I3f3ca55fb50efc503851c493aa4ccb279029e6d8
Reviewed-on: https://chromium-review.googlesource.com/1013451
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>