pygpt: Revise by adding partition number (NR) based API.
When talking about the partition entries, most programs will refer that
by 1-based "number of record (NR)". In PyGPT, partitions were stored in
a list and that makes accessing to partition in GPT requires a 0-based
index, and leaving lots of "number - 1" calculation everywhere.

As learned from image_tool, it is more natural to access partitions
using number (NR) so we are adding GetPartition to fetch partition
objects using 1-based number, and a new UpdatePartition that also looks
at number to find the location to update.
Also changed ExpandPartition to take number for input argument.

Note in GPT.Partition, there is a meta value for partition number so we
don't need to explicitly specify number again in UpdatePartition if we
want to update partition back to same location.

BUG=chromium:834237
TEST=make test

Change-Id: Ieceeb83183fffb315e0509b97cabe1f71efcb28e
Reviewed-on: https://chromium-review.googlesource.com/1017721
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>