goofy: don't reimplement upstart's shutdown procedures
All of these services have appropriate 'stop on' conditions, which get
triggered via the runlevel change that happens with the 'shutdown'
command ('shutdown -{h,r} now'). This list is just a red herring for
generating unnecessary cleanup.

If daemons aren't stopping when they should, modifications should be
made to their init files, not methods like this.

This is inspired by stuff like b/76121632 and CL:1022050.

BUG=chromium:835438
TEST=run factory toolkit reboot

Change-Id: I2a296fde5da27e56e467c8705674c5e300f4cae6
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1022019
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Yilun Lin <yllin@chromium.org>