probe: Fix bug in generic_usb_hosts function.
The super class of `generic_usb_hosts` function expects that the
method `ProbeDevice` returns a dictionary.  However, `generic_usb_hosts`
directly returns the probed results of `usb` and `pci` functions,
which must be a list which contains only one dictionary.  This CL
extracts that list and return the only element to solve the bug.

BUG=chromium:78441577
TEST=make test

Change-Id: I39bf1c69abc18597d195576fb4cd39ecc06cc440
Reviewed-on: https://chromium-review.googlesource.com/1023361
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Shen-En Shih <petershih@chromium.org>