gooftool: disable cr50 rma mode by looking into ccd info.
`gooftool -a -I` is supported in gsctool side now which can get ccd
info from the cr50. But currently the cr50 firmware (rw: 0.4.4) in tot
doesn't support it yet.

Before getting ccd info is ready in cr50 firmware (rw:0.4.5), we don't
know the fail of calling `gooftool -a -r disable` is normal (rma mode
is not enabled yet) or abnormal. Therefore what we can do is just try
call it in order to disable rma mode as we can as possible.

BUG=b:79113730
TEST=1) update cr50 version - 0.4.5 to the DUT.
2) check `gooftool cr50_disable_rma_mode`
3) change pattern from 0x1ffff to 0x16607 for simulating rma mode is
still enabled.
4) check `gooftool cr50_disable_rma_mode` again and see error msg.

Change-Id: I58d5dbece2b27133c177c51c2f17897e67d44591
Reviewed-on: https://chromium-review.googlesource.com/1041449
Commit-Ready: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>