hwid: Don't use default components if mismatched components are allowed.
Normally, the existance of a default component is a workaround so
that a bad probed values will not block the factory from generating
the HWID strings.  However, when we run hwid-updater, we consider the
probed values as a golden sample and encountering mismatched probed
values means that we need to record those values in the HWID database
so at this step, we should always ignore the default components.

BUG=chromium:845835
TEST=make test

Change-Id: I06918ba60aae310cd59b158e0324072f53bf792a
Reviewed-on: https://chromium-review.googlesource.com/1070831
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>