Dome: Fix deploy service config failed when not changed.
The default value of "active" is not set when the config didn't exist
initially, so when deploying a service config without clicking any form
element, and there wasn't a config initially for the service, the server
would get an empty dict, and would treat it as active=true, which is not
what we want.

Set the default value of active as whether the config exists to solve
this.

BUG=b:80217832
TEST=make test
TEST=manually, open dome, start an umpire instance, open the dkps
     service config section and click deploy, check that the deploy
     succeed.

Change-Id: Ia87567378c63085ab50eb187b8375a7ee7d0a296
Reviewed-on: https://chromium-review.googlesource.com/1080479
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>