Dome: Reduce polyfills for babel compiler.
There was lots of unnecessary polyfills for babel compiler, since the
>1% filter contains IE 11... This makes things hard to debug, in
particular, the compiled async functions has a really hard to understand
stack trace, and breakpoint can't be set on some lines. Since we
probably don't care if this works on IE or not (and we actually have
check to advise user to use Chrome), change the babel option to exclude
IE.

Also, be specific on what feature we need in preset-stage-2, and remove
the unused @babel/polyfill.

BUG=b:80217832
TEST=make test
TEST=make lint-frontend in py/dome
TEST=manually

Change-Id: I6fd92792a070b2d51ac2a7cdb7c2e96b386f4a77
Reviewed-on: https://chromium-review.googlesource.com/1080494
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>