Dome: Make task list title non-clickable.
It's easy to misclick on the task list title when the user want to click
on the "clear all waiting" or "dismiss all succeed" buttons, which cause
the task list to collapse. Change it so that only clicking on the "v" or
"^" button works.

Also move the collapsed state from DomeApp to TaskList.

BUG=b:80217832
TEST=make test
TEST=make lint-frontend in py/dome
TEST=manually.

Change-Id: I5c4c788afc4967906a040be3eab80ef89690846c
Reviewed-on: https://chromium-review.googlesource.com/1096805
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>