goofy_server: Fix the logic of handling files with unknown MIME type.
Fix the way GoofyWebRequestHandler checks if the MIME type of a file is
unknown, and change the behavior from rejecting the request with 404
file not found to setting mime type to 'application/octet-stream'.

Also make the server thread in GoofyServerTest a daemon thread, or the
unittest will block when Goofy server failed to get ready.

BUG=none
TEST=make test + manually run Goofy

Change-Id: Iedaed07b04c54515607fe578b9d2434f62113d82
Reviewed-on: https://chromium-review.googlesource.com/1122143
Commit-Ready: Youcheng Syu <youcheng@chromium.org>
Tested-by: Youcheng Syu <youcheng@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>