umpire: webapps: add a new webapp - Download Slots Manager
According to DUT (device under test) info embedded in request header, either it
will assign a new UUID to present a unique session if no uuid cookie presented
or keep a session alive mapped from the UUID in cookie. On the other hand, one
more information - N_PLACE will be returned which indicates whether the DUT gets
available download session (N_PLACE: 0) or been put in the queue due to slots
are all occupied already (ex: "N_PLACE: 3" shows that two DUTs are waiting for
the slot in front of it). In the end, the cookie of drop_slot would indicate
that this session is done so the slot can be reclaimed for other session.

BUG=b:80152439
TEST=manually test between DUT and factory server

Change-Id: Ic1822301d383f9360c1cc26ee4db408f591112be
Reviewed-on: https://chromium-review.googlesource.com/1127080
Commit-Ready: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>