Instalog: BufferEventStream doesn't time out in time when filtering
Instalog has flow policy to filter events in buffer.  If a output plugin
doesn't accept a event, the BufferEventStream will ignore it
automatically.  However, if buffer has lots of event which is not for
the output plugin, the BufferEventStream won't time out in time.

BUG=b:109905899
TEST=manually test in local device

Change-Id: Iccc9aad211fec47ecc80f9e591097303f30f5d90
Reviewed-on: https://chromium-review.googlesource.com/1141547
Commit-Ready: Chun-Tsen Kuo <chuntsen@chromium.org>
Tested-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>