yaml_utils: Decouples the YAML tag handler and the data class.
This CL modifies the helper classes in cros.factory.utils.yaml_utils
so that we don't need to set the meta-class of a data class to the
YAML tag handler.

Previously we need to define the YAML tag handler as the meta-class
of the data class.  So the handler must be defined "before" the data
class.  As the handler often knows some basic logic of the data class
to resolve the yaml data, these two classes become strongly coupled
and hards to be put in different files.

BUG=b:110596222
TEST=make test

Change-Id: I0a1e5c3ac572a5c655e6660abac136ccffe6469f
Reviewed-on: https://chromium-review.googlesource.com/1160078
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Yilun Lin <yllin@chromium.org>