regions: Don't trigger logger during import time.
When we run unittests in chroot, the `regions` module will execute
`platform2/regions/regions.py` to get both confirmed and unconfirmed
regions.  This causes an error when goofy configures the logger
during runtime.  This CL is just a shorterm solution because a better
solution should be preventing any side effects when the `regions`
module is imported.

BUG=b:110596222
TEST=make test

Change-Id: I464721a9dd3476945cd8b92994a0f4d938b6d360
Reviewed-on: https://chromium-review.googlesource.com/1175595
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Youcheng Syu <youcheng@chromium.org>