pytests/exec_shell: Optionally log the output of the commands.
The test could only upload the output files of the commands as
attachments.  However, attachment content cannot show on the
bigquery page directly so this CL adds an extra option to determine
whether the output of the commands should be uploaded by testlog or
not.

BUG=None
TEST=manaully test on DUT

Change-Id: I5918e08c3af66d388447506a765cce771137562c
Reviewed-on: https://chromium-review.googlesource.com/1179516
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>