Dome: Log task body when creating task.
Since we log all actions in console, it would make debugging easier if
we log task body when creating task too. To avoid the File object not
able to be garbage collected when it's no longer in use, we only put the
file name and file size, instead of the original File object.

BUG=b:80217832
TEST=make test
TEST=make lint-frontend in py/dome
TEST=manually.

Change-Id: Ibd47100ce7c1383696a7ae34213f373c164c5613
Reviewed-on: https://chromium-review.googlesource.com/1189529
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>