Dome: Add scrollbar to task list when there are many task.
When the task list is long, the task list would overflow the page,
the old tasks can't be seen, and the "dismiss all" button can't be
clicked. Make the task list scrollable when the task list is too long.

The buttons in the task list would be not aligned to the header in
current implementation, due to lack of support of CSS subgrid in
browsers.

BUG=b:80217832
TEST=make test
TEST=manually

Change-Id: Ibd6b8b2f9589c1cbe7dbdca73b2e5c13455ed0c7
Reviewed-on: https://chromium-review.googlesource.com/1189533
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>