hwid: Allow ambiguous components in HWID database.
Sometimes we realize that a volatile component value is recorded
by the HWID database and no longer want to use it as an identifier
value lately.  As the changes for HWID databases always need to
ensure the backward compatibility, the only solution is to treat
the component without that value as an individual component and
then record it in the HWID database.

BUG=None
TEST=make test; manually test on DUT

Change-Id: I5a088b243f56eaaee20fd4ce9633ee1c8051c32f
Reviewed-on: https://chromium-review.googlesource.com/1195191
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>