gooftool: add model name into the report name of GRT.
In the past, one umpire instance is to support one project / overlay.
But now, one umpire is to support one uni-build project which might
introduce multiple models. As a result, it is hard to identify which
model is this report for. In first level, we can solve it by
appending model name into the report name so partners are easy to identify
it.

BUG=b:113796166
TEST=manually test in the DUT with remote Umpire server.

Change-Id: I19c7f70352c7567a71002b68e8a0aa29a4aea984
Reviewed-on: https://chromium-review.googlesource.com/1203591
Commit-Ready: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>
Reviewed-by: Marco Chen <marcochen@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>