Dome: Add loading screen when checking auth token.
When Dome is checking if the auth token is valid, show a login screen
instead of the login form. This prevent the login form to be shown for
several seconds when there's no need to login again.
This also makes the frontend to wait for the backend server to start
before proceeding, since the first response for dev server / real server
takes about 3 and 1 seconds, respectively.

As a side effect, user can't no longer logout on localhost.

BUG=b:80217832
TEST=make test
TEST=manually

Change-Id: I01f0aa229ab6680ecd854d78f22adba159dedba5
Reviewed-on: https://chromium-review.googlesource.com/1206612
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@chromium.org>