Dome: Remove umpire_host.
Since Dome need to control Umpire using Docker, having Umpire instance
on other than localhost is never supported, and would not work. Remove
the field, and corresponding field in the frontend.

We can actually do better by auto-detecting if existing instance exists,
and use it directly if it is. This would be done in another CL.

BUG=b:80217832
TEST=make test
TEST=make test in py/dome
TEST=manually.

Change-Id: I889c94bcf826ab5693196b37770357bbffebaaf4
Reviewed-on: https://chromium-review.googlesource.com/1236034
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>