Dome: Derive types for props of redux-connected component.
Derive types for props of redux-connected components from
mapStateToProps and mapDispatchToProps, so we don't need to duplicate
those type definitions in two places. This makes it easier to change the
action signatures, without having to change those signatures in multiple
places.

The downside of this is that it's less clear to see what is available in
this.props in the component, but this should be easily overcomed by
using an IDE with typescript support.

Another upside for this is that the user of the component can see what
props need to be passed in more clearly, since the connected props are
separated from the "own" props.

Also update version for typescript to dev version, to avoid stepping
onto https://github.com/Microsoft/TypeScript/issues/25775.

BUG=b:80217832
TEST=make test
TEST=make lint-frontend in py/dome

Change-Id: Ie9888459863a7ff48823f852a9198ccff925d79c
Reviewed-on: https://chromium-review.googlesource.com/1248326
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>