pytests/summary: Refactor the code to render the table by JS.
The original implementation rendered all the test UI in the python
part, which includes a table for the test results.  That
implementation was not easy to read and maintain because the UI
of this pytest is not very simple.  Therefore, this CL implements a
JS frontend class to handle the UI related works.

BUG=b:118277648
TEST=manually test on DUT

Change-Id: Ibc4637d39ee6f722f026e137b31c4081c6544d76
Reviewed-on: https://chromium-review.googlesource.com/1297801
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>