pytests/summary: Implement a screensaver.
Sometimes we don't want the display to show a static pattern too
long.  And because the toolkit often stops at the `summary` pytest
at the end of a series of tests, this CL let the pytest enter a
screensaver which color will change with time.

BUG=b:118277648
TEST=manually test on DUT

Change-Id: Ia4bd378e5774c5b7f8117371df30bbb3e47a376e
Reviewed-on: https://chromium-review.googlesource.com/1297802
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>