cutoff.sh: To prevent a system crash on Factory Shim.
When we going to cutoff device on Factory Shim, it may reboot if the
battery percentage is higher than "CUTOFF_BATTERY_MAX_PERCENTAGE". This
is because stressapptest run out of memory of the device. This CL limit
the memory usage of stressapptest, and it also speed up the process of
discharging.

BUG=b:117961220
TEST=manually test on local device

Change-Id: Ie5971be582c53548c121cd8177a968e83efb28e2
Reviewed-on: https://chromium-review.googlesource.com/1307254
Commit-Ready: Chun-Tsen Kuo <chuntsen@chromium.org>
Tested-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>