goofy: Initialize state_instance before loading the test lists.
CL:1249181 moved the initialization of the state_instance later
than the test list initialization.  It was because Goofy might have
to modify the shared data after it resolved the correct test list ID.
However, the test list initialization requires the state_instance to
be ready because it saves the failure reason to the shared data so
this CL refactors the code to solve the chicken-egg problem.

BUG=None
TEST=manual test; make test

Change-Id: I861c808c6c87c55d7e236fc4fffbdc0abc8630be
Reviewed-on: https://chromium-review.googlesource.com/1326162
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Marco Chen <marcochen@chromium.org>