registration code: don't allow dummy registration code in PVT only
CL:1276026 inhibit dummy regcode in all phases.  This is not very
convenient for early builds.  Allow this code in non-PVT builds.
If dummy code is detected, and it's allowed in current phase, a warning
will be logged.  If it's not allowed, an exception will be thrown.

Also add argument '--allow_dummy' to regcode.py.

BUG=b:117463731
TEST=manual

Change-Id: I6ec60754c7ca3f53c3630f1ff2bf269e1b7b3ae2
Reviewed-on: https://chromium-review.googlesource.com/1398101
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>