config_utils: fix loading json config in factory.par
The config_utils.py will fail if a config file doesn't exist in PAR.
We should always ignore this error in _LoadJsonFile function, and let
caller to decide if missing file is a critical failure.

Also add an unittest to catch this kind of issue in the future.

BUG=b:124619303
TEST=make test
TEST=manual test

Change-Id: I37336759a50f50320c68571d83f84a7871084786
Reviewed-on: https://chromium-review.googlesource.com/1477133
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>