pytest_invoker: Report structured failure reason.
This CL revises the related modules to record the exception from
the pytest and report to the invoker without turning it into a
string of error message.  As the test result from the test case
becomes structured, `goofy` can parse them easier to provide
features like showing the line of the source code that raises
the exception.

BUG=chromium:933275
TEST=make test

Change-Id: Id443e4caf792cb8e17ff081944b3863e3a3ab176
Reviewed-on: https://chromium-review.googlesource.com/1477132
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>