gooftool: wipe: don't delete .developer_mode if GBB flags are not cleared
In early build, we might decide to keep GBB flags so we can stay in
developer mode. In this case, clobber-state will still generate
'.developer_mode' tag in stateful partition, and wipe will still delete
it.  So on next boot (in release image), Chrome OS can't find the tag,
and will start power washing immediately, which is time consuming.  To
prevent this, add code to detect if GBB flags are cleared, if not and
dev switch is forced on, we won't delete the tag.

BUG=None
TEST=make test

Change-Id: Ibef294472ffb44f1e38d5606151e6ea00432a623
Reviewed-on: https://chromium-review.googlesource.com/1485561
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>