pytests/update_cr50_firmware: Provide test flow to verify the version.
The original test flow to update the cr50 firmware is to move
the new firmware image into section B and reboot.  The test
passes if the updater succeeds.  However, after DUT reboots,
the firmware might fail to switch to the new firmware so
we should also check the version then.

Consider that the test flow for updating and checking the
firmware are quite similar, the CL implements them in the same
pytest which provides an argument to switch the goal.

Also, this CL extracts the code that invokes `gsctool` command
line utility into an individual helper class to prevent the
duplicated logic.

BUG=b:127906738
TEST=manual test

Change-Id: Ifae16dae8e5655a2c6a577fb00a5e4b742aa8299
Reviewed-on: https://chromium-review.googlesource.com/1520548
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>