hwid: add new status 'duplicate'
In CL:1195191, we have covered the case that a component is subset of
other component, if regular expression is not used.  This should have
covered most cases.

To cover the regular expression case, for example:

  comp_1:
    values:
      part_number: PART_NUMBER_1-1
  comp_2:
    values:
      part_number: PART_NUMBER_1-2

And we later found that, PART_NUMBER_1-1 and PART_NUMBER_1-2 are
actually the same component in AVL database, 'PART_NUMBER_1', with some
suffix ('-1' and '-2') that we actually don't care.  The following
change [1] should not work:

  comp_1:
    values:
      part_number: !re PART_NUMBER_1-.*
  comp_2:
    values:
      part_number: PART_NUMBER_1-2

Because comp_2 is a subset of comp_1.  (There is a bug in current code,
so the encoder won't complain, but it is wrong).  To solve this kind of
ambiguity, status 'duplicate' is introducted.

  comp_1:
    values:
      part_number: !re PART_NUMBER_1-.*
  comp_2:
    status: duplicate
    values:
      part_number: PART_NUMBER_1-2

In this case, encoder will ignore comp_2.  And comp_2 can still be used
for decoding.  Also fix the bug in cros.factory.hwid.v3.probe, so the
change like [1] will fail during HWID encode.

BUG=None
TEST=make test

Change-Id: Ibc8938d0308930d5fbb23d138fdf117e46f0b6b8
Reviewed-on: https://chromium-review.googlesource.com/1535649
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Yong Hong <yhong@google.com>