bad_blocks: ignore FAILERR and FAILLOG
bad_blocks test in factory should be aligned with autotest
hardware_Smartctl.  The autotest ignores FAILERR and FAILLOG when
determining if the test passed or not.

In this change, we modify the bad_blocks test to align with the
autotest.

BUG=b:130521180
TEST=None

Change-Id: Ie8eda6fac1e9cfbc0234af709ba34e0accb85f8e
Reviewed-on: https://chromium-review.googlesource.com/1577446
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Marco Chen <marcochen@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>