py/device: Support motion sensor setup for gyroscope
We used to create a GyroscopeGroup for a complete gyroscope test cycle,
which includes:
- motion sensor setup
- calibration
- main gyro test

Setting sensor is mandatory before running any gyroscope relating test,
but the setup command is not properly supported and we need to add it
manually into test list, which makes things tedious and error prone.

In addition, though ectool provides gyroscope information, we were
not leveraging that to provide functionalities e.g. parameter checking
or auto adaption.

This CL tries to add some supports for that in gyroscope.py so pytests
can call SetupMotionSensor() to set it up, and if the needed parameters
like ID or frequency are not provided as arguments, help functions
will collect the information from ectool and try to automatically fill
the blanks up.

BUG=b:133213368
TEST=make toolkit BOARD=${BOARD}; tested on Kukui

Change-Id: If90bce6dbe7692f0146551f052cefd83c6d01222
Reviewed-on: https://chromium-review.googlesource.com/1627173
Tested-by: Fei Shao <fshao@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>