factory: Use list(dict.values()) when not iterating
In python2, `dict.values()` return a copy list of values.
In python3, `dict.values()` return an iterator of values.

The semantic changes, so we need to use `list()` to wrap it to get a copy
of list when it's not the case of iterating.

BUG=chromium:999876
TEST=make test
TEST=make lint

Change-Id: I664ad896c5a5b42bd093481b45f1138cdb34e7a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1875829
Tested-by: Yilin Yang <kerker@chromium.org>
Auto-Submit: Yilin Yang <kerker@chromium.org>
Commit-Queue: Yilin Yang <kerker@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>