factory: Use list(dict.items()) when not iterating
In python2, `dict.items()` return a copy list of items.
In python3, `dict.items()` return an iterator of items.

The semantic changes, so we need to use `list()` to wrap it to get a copy
of list when it's not the case of iterating.

BUG=chromium:999876
TEST=make test
TEST=make lint

Change-Id: I5db9ceb3a3ccdba700383e0c64fece0a657cb0a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1875830
Tested-by: Yilin Yang <kerker@chromium.org>
Auto-Submit: Yilin Yang <kerker@chromium.org>
Commit-Queue: Yilin Yang <kerker@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>