factory: Import mox from mox3 module
`mox` is not supported in python3. The formal way to solve this issue
is to use `mock` module. But there are too many files that need to be
migrated, so we use `mox3` as a temporary solution.

We need to migrate them into `mock` module in future, keep track on
b:147990582.

BUG=chromium:999876, b:147990582
TEST=None

Change-Id: If3ff5a43e37ca04f749913b7b9af651ee8f541c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1991205
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>