goofy: Make iterations of test objects mutable in engineering mode.
When a test failed, we often want to retry multiple times to see the
fail rate. It would be useful to enter engineering mode and set how many
times we want to retry the test in the UI. This CL makes user can right
click the test and then choose "set the test to run multiple times".
Finally, a box pops up and the user inputs the time of retries.

In order to keep the modification through RebootStep, this CL saves
iterations and retries to TestState.

BUG=b:69115897
TEST=(1) Set multiple tests to run multiple times and run those tests.
     (2) Run RebootStep and the modification of other tests is kept.
     (3) Set RebootStep to run multiple times and run it.
     (4) Reload test list or clear test state to recover.

Change-Id: I324338480a11a0988c496059f574cd5290f5f745
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1932871
Tested-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>