hwid/v3: improve HWIDv3 database test coverage
Because previously HWID database files are located in different
branches, it's hard to detect if the file to test should be loaded
from a remote branch or from local disk.  As today all files are
located in the master branch, this CL leverages the presumption
to load the correct file to test to increase the coverage.

TEST=make test
BUG=None

Change-Id: Ib0c55111fd54f7f6c1f23ff3cb87762e16319a4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1943499
Reviewed-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>