instalog: Use BytesIO / StringIO properly
According to /usr/lib64/python3.6/cgi.py, it accept `str` and `bytes`,
and we can use `_binary_file` to determine the type. If it's true, then
we need to use `BytesIO` and open file in 'wb' mode. If not, then we
need to use `StringIO` and open file in 'w' mode.

BUG=chromium:999876
TEST=make test and realted unittest pass

Change-Id: I9aae8cde3d9aa702c942c4cf02b4ef0af84db4d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1991249
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>