factory: Copy dict items if we need to delete items when iterating
It is abandoned when we iterate a dict through `dict.items()`,
`dict.keys()`, `dict.values()` and delete items in the meantime.
If we want to do this, we need to copy the dict items and iterate on the
copy list.

BUG=chromium:999876
TEST=None

Change-Id: Iccaca026570c5f4650b5345f0cb531188e7bc42f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1991255
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Auto-Submit: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>