bad_blocks: Store locals() in a variable
When using `locals()['key']` directly, it will cause
`TypeError("'builtin_function_or_method' object is not subscriptable",)`,
because python treats it as a function not a list.

So we need to store it in a variable, and then access it.

BUG=chromium:999876
TEST=manual test on dut and the error disappear

Change-Id: I20b10af95972f555b7c07d116f61caa879f5dc25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1991282
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Auto-Submit: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>