probe_info_service: setup unittest and manual test flow
The working environment of Probe Info Service is quite different
than the factory framework on DUT so the default routine to run the
unittests is not suitable for this project.  Therefore, this CL
implements a standalone test runner for Probe Info Service.

In addition, this CL enables the feature of running the instance
locally to make manual test more friendly.

BUG=b:146768256
TEST=manual test

Change-Id: Ia3f1b0b1fbab403ae5f963dd22fa9ef9864af958
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1992968
Reviewed-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>