service_utils: Allow ServiceManager handle multiple setup/restore requests
ServiceManager is designed to serve pytests or sequential test groups,
and enabling/disabling services is not supported for parallel
test groups: assigning service list is ineffective to parallel groups
and invalid to the subtests. But in some scenarios, we do want to create
an environment for parallel test group with certain services enabled or
disabled e.g. run-in stressing.

This CL modified the implementation of ServiceManager to handle several
setup/restore calls as long as they are in valid sequence (think of
parentheses matching).

This doesn't affect the usage of singular tests because there's only one
setup and restore. For the case of parallel groups, the service lists
are the same across subtests, so the setup and restore requests can be
called in arbitrary order and only the first setup and last restore take
effect.

Also fixed a lint warning and docstring quotings.

BUG=b:146626823
TEST=manual test, set "disable_services" to a parallel group and see
corresponding services suspended/resumed during/after group test
execution

Change-Id: If3e4a439f491d26353a181642a2afda8e0a98c50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1980601
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>