rf_graphyte: Fix serial_number becoming bytes in py3
In python3, the old method to get serial number returns bytes and the
name of the log file becomes b'XXXX'. This CL changes it to use
device_data so we can get the string version of serail number.

This CL works in py2 as well.

BUG=chromium:707200
TEST=Run rf_graphyte on devices.

Change-Id: I503e7165df2dc5f2f325194b0b886a3494c2387a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2035434
Reviewed-by: Yong Hong <yhong@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>