hwid: fix potential indexing error in region field
While dumping the data to HWID databases, the original implementation
outputs the return elements of `dict.values()` as a list of regions
in region field directly.  However, the method doesn't guarantee the
order of the elements, which makes the code works just out of good
fortune.

BUG=None
TEST=make test

Change-Id: I28e155ae55e9c405637f6e6dbcbe86e548b9b7fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2046643
Reviewed-by: Clark Chung <ckclark@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>