gooftool: fix incorrect symlink in testdata
The testdata folder contains a golden HWID database for gooftool
instance to load in the unittest routine.  Since the unittest doesn't
actually tests the content of the database, that HWID database is
merely a symbolic link to the sample DB in `py/hwid/v3/testdata`
to re-use the content.  While the target file were renamed, this
symlink was forgotten to be update as well.

This CL fixes the symlink destination, as well as adding a test to
cover the case.

BUG=None
TEST=None

Change-Id: I6631213488bf21fbcbd595b9a314cfe3eb2fc48f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2083071
Tested-by: Yong Hong <yhong@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Auto-Submit: Yong Hong <yhong@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>