update_sku: allows different fw-config between EEPROM and model.yaml
In the factory SMT stage, the fw-config in EEPROM may be preflashed to
a value which meets the SMT fixture requirement. However, the fw-config
preflashed in the CBI EEPROM may not same as the model.yaml because
the fw-config value may change after the device assembles (for example
different USB DTB). This patch adds a check: if the sku id isn't equal
to the default sku id(ie.0x7fffffff), we will block the fw-config to
be programmed to the DUT when the CBI EEPROM is inconsistent with
cros_config.

BUG=none
TEST=build pass

Change-Id: Ie3dee1b236f3f3553a3a5dcdf818748bcdbbcfd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2107050
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Auto-Submit: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>