goofy plugin: add led_indicator plugin
led_indicator was a pytest in fizz that changes the color of LED to
indicate test state of specified tests. Moved this to public repo for
that this is useful for multiple boards. Implemented this as a plugin
to regularly check test state and change LED status.

BUG=b:150475702
TEST=manually test on puff

Change-Id: I4b3d851c1abc0a1308e68147bbb73ad4da1af0e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2106049
Reviewed-by: Stimim Chen <stimim@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>