service_utils: Ignore disabling non-existing service.
Previously, when we try to disable a non-existing service, the
invocation throws an exception and does not fail or continue the test.
This makes the toolkit not continue running on the next test.

This CL adds changes so Goofy ignores services which throw exceptions
while disabling or enabling.

BUG=None
TEST=Run a test which disables non-existing services and it completes.

Change-Id: I92131946375471b9f1e4a4b75507a1dbcf5ed91f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2120766
Reviewed-by: Stimim Chen <stimim@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>